SQL schema: add columns for referencing IDN and ACE versions
Postfix with SMTPUTF8 support expects to find domains names and email
addresses for IDN domains in both versions, as UTF8 and as ACE string.
The new columns make it possible to have a relationship between this
two representations. Skaffari will only show the UTF8 version in the
frontend but will maintain the ACE versions as shadows in the datbase so
that Postfix can find them.